refactor(storybook): restructure file locations, theme loading
