+++ docs/addons.js
@@ -1,3 +1,2 @@
 import '@storybook/addon-actions/register'
 import '@storybook/addon-links/register'
-import 'storybook-addon-styled-component-theme/dist/src/register' // v1.1.0^
