+++ package.json
@@ -120,7 +120,6 @@
     "react-window": "^1.5.2",
     "rimraf": "^2.6.2",
     "standard-version": "^4.4.0",
-    "storybook-addon-styled-component-theme": "^1.1.1",
     "tslint": "^5.12.1",
     "tslint-config-standard": "^8.0.1",
     "typescript": "^3.3.3"
