--- src/fontConfig.ts
@@ -1,8 +1,0 @@
-import './assets/styles/fonts.css'
-
-export const fontConfig = {
-  custom: {
-    families: ['Fira Mono Latin Regular', 'Pt Sans Latin Bold', 'Pt Sans Latin Regular'],
-    urls: ['./assets/styles/fonts.css']
-  }
-}
