+++ src/utils/testing/decorators/ThemeDecorator.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react'
+import ThemeAndFontProvider from '../../../ThemeProvider'
+
+export const ThemeDecorator = (storyFn: any) => (
+  <ThemeAndFontProvider>
+    {storyFn()}
+  </ThemeAndFontProvider>
+)
